refactor: 데이터 유지와 관련된 코드들 새로운 클래스로 합침
