+++ src/index.js
@@ -1,16 +1,11 @@
 import { VendingMachineHandler } from "./vendingMachineHandler.js";
 import { Product } from "./models/product.js";
 import { Coin } from "./models/coin.js";
-import { LocalDataPersister } from "./localDataPersister.js";
-import {
-    convertListToCoinsObj,
-    convertListToProductObj,
-    convertListToUserBalanceObj
-} from "./utils/utils.js";
+import { VendingMachinePersister } from "./persisters/vendingMachinePersister.js";
 
 const isDevelop = false;
 
-const dataPersister = new LocalDataPersister();
+const vendingMachinePersister = new VendingMachinePersister();
 
 const mockData = {
     products: [
@@ -29,15 +24,7 @@ const mockData = {
     }
 }
 
-const productData = dataPersister.getData('products');
-const machineCoinsData = dataPersister.getData('machineCoins');
-const userBalanceData = dataPersister.getData('userBalance');
-
-const deployData = {
-    products: convertListToProductObj(productData),
-    machineCoins: convertListToCoinsObj(machineCoinsData),
-    userBalance:convertListToUserBalanceObj(userBalanceData)
-}
+const deployData = vendingMachinePersister.getVendingMachineData();
 
 const initialData = isDevelop ? mockData : deployData;
 const vendingMachine = new VendingMachineHandler(initialData);
